md5: add RFC1321-based (RSA-free) MD5 library
Obtained from http://libmd5-rfc.sourceforge.net/:
This is a very small C library implementing RFC1321, the MD5 message
digest function. Unlike the existing W3C libmd5, it was written from the
specifications (not the sample code) in RFC1321, and therefore is not
required to acknowledge RSA in any way.